+++ src/stac_auth_proxy/filters/__init__.py
@@ -1,4 +1,4 @@
-"""CQL2 filter generators."""
+"""CQL2 filter factories."""
 
 from .opa import Opa
 from .template import Template