+++ src/stac_auth_proxy/handlers/reverse_proxy.py
@@ -9,6 +9,8 @@ from fastapi import Request
 from starlette.datastructures import MutableHeaders
 from starlette.responses import Response
 
+from stac_auth_proxy.utils.requests import build_server_timing_header
+
 logger = logging.getLogger(__name__)
 
 
@@ -37,14 +39,18 @@ class ReverseProxyHandler:
         headers = MutableHeaders(request.headers)
         headers.setdefault("Via", f"1.1 {self.proxy_name}")
 
-        proxy_client = request.client.host if request.client else "unknown"
-        proxy_proto = request.url.scheme
-        proxy_host = request.url.netloc
-        proxy_path = request.base_url.path
+        proxy_client = headers.get(
+            "X-Forwarded-For", request.client.host if request.client else "unknown"
+        )
+        proxy_proto = headers.get("X-Forwarded-Proto", request.url.scheme)
+        proxy_host = headers.get("X-Forwarded-Host", request.url.netloc)
+        proxy_path = headers.get("X-Forwarded-Path", request.base_url.path)
         headers.setdefault(
             "Forwarded",
             f"for={proxy_client};host={proxy_host};proto={proxy_proto};path={proxy_path}",
         )
+
+        # NOTE: This is useful if the upstream API does not support the Forwarded header
         if self.legacy_forwarded_headers:
             headers.setdefault("X-Forwarded-For", proxy_client)
             headers.setdefault("X-Forwarded-Host", proxy_host)
@@ -82,11 +88,15 @@ class ReverseProxyHandler:
         start_time = time.perf_counter()
         rp_resp = await self.client.send(rp_req, stream=True)
         proxy_time = time.perf_counter() - start_time
-
+        rp_resp.headers["Server-Timing"] = build_server_timing_header(
+            rp_resp.headers.get("Server-Timing"),
+            name="upstream",
+            dur=proxy_time,
+            desc="Upstream processing time",
+        )
         logger.debug(
             f"Received response status {rp_resp.status_code!r} from {rp_req.url} in {proxy_time:.3f}s"
         )
-        rp_resp.headers["X-Upstream-Time"] = f"{proxy_time:.3f}"
 
         # We read the content here to make use of HTTPX's decompression, ensuring we have
         # non-compressed content for the middleware to work with.