+++ src/stac_auth_proxy/middleware/Cql2BuildFilterMiddleware.py
@@ -93,7 +93,7 @@ class Cql2BuildFilterMiddleware:
             cql2_filter.validate()
         except ValidationError:
             logger.error("Invalid CQL2 filter: %s", filter_expr)
-            return await Response(status_code=502, content="Invalid CQL2 filter")
+            return Response(status_code=502, content="Invalid CQL2 filter")
         setattr(request.state, self.state_key, cql2_filter)
 
         return await self.app(scope, receive, send)